+++ packages/api/src/lib/inferenceNodeApi.ts
@@ -2,5 +2,5 @@ export { commit } from "./inferenceNodeApi/commit.js";
 export { create } from "./inferenceNodeApi/create.js";
 export { decode } from "./inferenceNodeApi/decode.js";
 export { delete_ as delete } from "./inferenceNodeApi/delete.js";
-export { infer } from "./inferenceNodeApi/infer.js";
+export { infer, inferStream } from "./inferenceNodeApi/infer.js";
 export { tokenCount } from "./inferenceNodeApi/tokenCount.js";