+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 
 - [x] feat: inference server (Sat June 15)
 - [x] feat: GPT session caching (Wed June 19)
+- [x] feat: streaming inference (Sun June 23)
 
 ### May 2024